Refactored config utils
