+++ tsdat/config/utils/_named_class.py
@@ -1,0 +1,7 @@
+from typing import (
+    Protocol,
+)
+
+
+class _NamedClass(Protocol):
+    name: str
