+++ tsdat/config/utils/config.py
@@ -1,0 +1,9 @@
+from typing import (
+    TypeVar,
+)
+from jsonpointer import set_pointer  # type: ignore
+from pydantic import (
+    BaseModel,
+)
+
+Config = TypeVar("Config", bound=BaseModel)
