+++ tsdat/config/utils/config_error.py
@@ -1,0 +1,2 @@
+class ConfigError(Exception):
+    pass
