+++ tsdat/config/utils/matches_overrideable_schema.py
@@ -1,0 +1,8 @@
+from typing import (
+    Any,
+    Dict,
+)
+
+
+def matches_overrideable_schema(model_dict: Dict[str, Any]):
+    return "path" in model_dict
